Agregada ruta de Resumen con el conteo de documentos para cada colección, en la versión de MongoDB

+++ MICHIPEDIA_CS_REST_NoSQL_API/MICHIPEDIA_CS_REST_NoSQL_API/Models/Resumen.cs
@@ -5,15 +5,15 @@ namespace MICHIPEDIA_CS_REST_NoSQL_API.Models
     public class Resumen
     {
         [JsonPropertyName("razas")]
-        public int Razas { get; set; } = 0;
+        public long Razas { get; set; } = 0;
 
         [JsonPropertyName("paises")]
-        public int Paises { get; set; } = 0;
+        public long Paises { get; set; } = 0;
 
         [JsonPropertyName("caracteristicas")]
-        public int Caracteristicas { get; set; } = 0;
+        public long Caracteristicas { get; set; } = 0;
 
         [JsonPropertyName("comportamientos")]
-        public int Comportamientos { get; set; } = 0;
+        public long Comportamientos { get; set; } = 0;
     }
 }
